Introduce script::ScriptExt
In preparation for moving the `Script` type to `primitives` introduce a
`ScriptExt` trait using the `define_extension_trait` macro.

Note that we convert private methods to stand alone functions instead
of creating an additional private trait.
